+++ changes.md
@@ -4,6 +4,7 @@
 1. Fully avoid jQuery on starting GUI, load only for menu or special displays
 2. Provide bootstrap-based menu - optionally
 3. Make "col" default draw option for TH2 in default JSROOT GUI
+4. Upgrade d3.js to 7.1.1, take only used components: "d3-selection", "d3-drag", "d3-scale"
 
 
 ## Changes in 6.3.0