+++ changes.md
@@ -9,6 +9,7 @@
 6. Implement proper drawing of TEllipse
 7. Support new TGraphMultiErrors class
 8. Implement all variants of CANDLE and VIOLIN draw options (#194)
+9. Implement "cjust" draw option when drawing color palette
 
 
 ## Changes in 6.3.0